Refactor gigahorse.py to allow different fact generators. WIP.
