+++ src/common.py
@@ -1,5 +1,15 @@
 from os.path import abspath, dirname, join, exists
 
+
+GIGAHORSE_DIR = join(dirname(abspath(__file__)), '..')
+"""The path of the gigahorse-toolchain clone."""
+
+DEFAULT_SOUFFLE_BIN = 'souffle'
+"""Location of the Souffle binary."""
+
+SOUFFLE_COMPILED_SUFFIX = '_compiled'
+
+
 def __get_sig_file(simple_filename):
     preferred_dest = join(join(dirname(abspath(__file__)), '../../common-facts'), simple_filename)
     fallback_dest = join(join(dirname(abspath(__file__)), '..'), simple_filename)
